+++ tests/virt/pipe.rs
@@ -369,7 +369,7 @@ impl<'a, const N: usize> Pipe<'a, N> {
                 }
                 match self.interchange.request((
                     request.command,
-                    Bytes::from_slice(&self.buffer[..request.length as usize]).unwrap(),
+                    Bytes::try_from(&self.buffer[..request.length as usize]).unwrap(),
                 )) {
                     Ok(_) => {
                         self.state = State::WaitingOnAuthenticator(request);